Jwang

+++ Assets/topoScene/WWWHandler.cs
@@ -50,7 +50,7 @@ public class WWWHandler : MonoBehaviour {
 	const string accesKeyMapBox = ".json?radius=5000&access_token=pk.eyJ1IjoiamN0d2FrZSIsImEiOiJjamQ1NHN2MGEweDJkMndxcmI3eHRuczRlIn0.if6fE47kjlJQbrKmRMMpZg";
 
 
-	public IEnumerator PostAnnotation(string signType, string text, float lat, float lon)
+	public IEnumerator PostAnnotation(string signType, string text, double lat, double lon)
 	{
 		WWWForm form = new WWWForm();
 		form.AddField("type", signType);
@@ -143,7 +143,6 @@ public class WWWHandler : MonoBehaviour {
 	}
 
 	public IEnumerator GetTrail(string name){
-
 		if (!String.IsNullOrEmpty (name)) {
 			//string isn't null, try to query
 			StringBuilder query = new StringBuilder (getTrailUrl);
